refactor(ui): extract browser notifications into dedicated hook and helper
